+++ src/main/java/com/tw/ticket/dao/CouponRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.tw.ticket.model.Coupon;
 
 @Repository
-public interface CouponRepository extends JpaRepository<Coupon, Long> {
+public interface CouponRepository extends JpaRepository<Coupon, Integer> {
 	//
 }
