+++ src/main/java/com/tw/ticket/dao/TicketCartRepository.java
@@ -14,8 +14,8 @@ public interface TicketCartRepository extends JpaRepository<TicketCart, PrimaryK
 	// 不須實作 JpaRepository會自動解析 (其中 findBy And 是關鍵字)
 	// KeyMemberId 表示 是變數key內的 memberId
 	// KeyTicketId 表示 是變數key內的 ticketId
-	public TicketCart findByKeyMemberIdAndKeyTicketId(long memberId, long ticketId);
+	public TicketCart findByKeyMemberIdAndKeyTicketId(int memberId, int ticketId);
 
 	// 找出同一個MemberId的TicketCart
-	public List<TicketCart> findByKeyMemberId(long memberId);
+	public List<TicketCart> findByKeyMemberId(int memberId);
 }
