+++ src/main/java/com/tw/ticket/dao/TicketCommentRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 import com.tw.ticket.model.TicketComment;
 
 @Repository
-public interface TicketCommentRepository extends JpaRepository<TicketComment, Long> {
+public interface TicketCommentRepository extends JpaRepository<TicketComment, Integer> {
 
 	// 取得指定票券編號的評論
-	public Page<TicketComment> findAllByTicketId(final long ticketId, Pageable pageable);
+	public Page<TicketComment> findAllByTicketId(final int ticketId, Pageable pageable);
 }
