+++ src/main/java/com/tw/ticket/dao/TicketImageRepository.java
@@ -1,15 +1,11 @@
 package com.tw.ticket.dao;
 
-
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.tw.ticket.model.TicketImage;
 
 @Repository
-public interface TicketImageRepository extends JpaRepository<TicketImage, Long> {
-
-
-
+public interface TicketImageRepository extends JpaRepository<TicketImage, Integer> {
+	//
 }
