+++ src/main/java/com/tw/ticket/dao/TicketSnRepository.java
@@ -10,9 +10,9 @@ import org.springframework.stereotype.Repository;
 import com.tw.ticket.model.TicketSn;
 
 @Repository
-public interface TicketSnRepository extends JpaRepository<TicketSn, Long> {
+public interface TicketSnRepository extends JpaRepository<TicketSn, Integer> {
 
 	// 關鍵字搜尋 hql
 	@Query("SELECT t FROM TicketSn t WHERE t.ticket.ticketId =:id AND t.status=0")
-	public List<TicketSn> searchUsableSn(@Param("id") long id);
+	public List<TicketSn> searchUsableSn(@Param("id") int id);
 }
