+++ src/main/java/com/tw/ticket/model/Coupon.java
@@ -20,7 +20,7 @@ public class Coupon implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long couponId;// 優惠券編號
+	private Integer couponId;// 優惠券編號
 
 	private String name;// 優惠券名稱
