+++ src/main/java/com/tw/ticket/service/ImageService.java
@@ -3,5 +3,5 @@ package com.tw.ticket.service;
 public interface ImageService {
 
 	// 找出指定圖片
-	public byte[] findImg(final long id);
+	public byte[] findImg(final int id);
 }
